Standalone changes for es5-persistence: Updated classpaths to remove versioning; Updated README.md to reflect latest instructions.
