+++ es5-persistence/README.md
@@ -1,6 +1,6 @@
 ## Usage
 
-1. In `server/build.gradle` file,  add `compile project(':conductor-es5-persistence')` in dependencies
-1. In `server/ConductorServer.java` file , replace  `import com.netflix.conductor.dao.es.EmbeddedElasticSearch` with `import com.netflix.conductor.dao.es5.es.EmbeddedElasticSearch`
-1. In `server/ServerModule.java` file,  replace  `import com.netflix.conductor.dao.index.ElasticSearchDAO; import com.netflix.conductor.dao.index.ElasticsearchModule` with `import com.netflix.conductor.dao.es5.index.ElasticSearchDAO;  import com.netflix.conductor.dao.es5.index.ElasticsearchModule;`
-1. Config property 'workflow.elasticsearch.cluster.name' , value with your elasticsearch cluster name
+1. In `server/build.gradle` file,  change to `compile project(':conductor-es5-persistence')` in dependencies, if not present already.
+2. In `settings.gradle` file, change to `es5-persistence`, if not present already.
+Make sure only one implementation of Elasticsearch (i.e es{v}-persistence) exists in gradle dependency tree.
+3. Generate new locks with `gradlew generateLocks saveLocks`.
