+++ es5-persistence/build.gradle
@@ -8,7 +8,6 @@ dependencies {
 	compile "org.apache.logging.log4j:log4j-core:${revLog4jCore}"
 
 	compile "org.elasticsearch:elasticsearch:${revElasticSearch5}"
-	compile "org.elasticsearch.client:transport:${revElasticSearch5Client}"
+	compile "org.elasticsearch.client:transport:${revElasticSearch5}"
 	compile "commons-io:commons-io:${revCommonsIo}"
-
 }
