+++ es5-persistence/dependencies.lock
@@ -56,7 +56,7 @@
             "firstLevelTransitive": [
                 "com.netflix.conductor:conductor-core"
             ],
-            "locked": "0.68.0"
+            "locked": "0.40.0"
         },
         "commons-io:commons-io": {
             "locked": "2.4",
@@ -154,7 +154,7 @@
             "firstLevelTransitive": [
                 "com.netflix.conductor:conductor-core"
             ],
-            "locked": "0.68.0"
+            "locked": "0.40.0"
         },
         "commons-io:commons-io": {
             "locked": "2.4",
@@ -252,7 +252,7 @@
             "firstLevelTransitive": [
                 "com.netflix.conductor:conductor-core"
             ],
-            "locked": "0.68.0"
+            "locked": "0.40.0"
         },
         "commons-io:commons-io": {
             "locked": "2.4",
@@ -350,7 +350,7 @@
             "firstLevelTransitive": [
                 "com.netflix.conductor:conductor-core"
             ],
-            "locked": "0.68.0"
+            "locked": "0.40.0"
         },
         "commons-io:commons-io": {
             "locked": "2.4",
@@ -448,7 +448,7 @@
             "firstLevelTransitive": [
                 "com.netflix.conductor:conductor-core"
             ],
-            "locked": "0.68.0"
+            "locked": "0.40.0"
         },
         "commons-io:commons-io": {
             "locked": "2.4",
@@ -546,7 +546,7 @@
             "firstLevelTransitive": [
                 "com.netflix.conductor:conductor-core"
             ],
-            "locked": "0.68.0"
+            "locked": "0.40.0"
         },
         "commons-io:commons-io": {
             "locked": "2.4",
@@ -652,7 +652,7 @@
             "firstLevelTransitive": [
                 "com.netflix.conductor:conductor-core"
             ],
-            "locked": "0.68.0"
+            "locked": "0.40.0"
         },
         "commons-io:commons-io": {
             "locked": "2.4",
@@ -758,7 +758,7 @@
             "firstLevelTransitive": [
                 "com.netflix.conductor:conductor-core"
             ],
-            "locked": "0.68.0"
+            "locked": "0.40.0"
         },
         "commons-io:commons-io": {
             "locked": "2.4",
@@ -864,7 +864,7 @@
             "firstLevelTransitive": [
                 "com.netflix.conductor:conductor-core"
             ],
-            "locked": "0.68.0"
+            "locked": "0.40.0"
         },
         "commons-io:commons-io": {
             "locked": "2.4",
