+++ es5-persistence/src/main/java/com/netflix/conductor/dao/es/index/query/parser/AbstractNode.java
@@ -16,7 +16,7 @@
 /**
  * 
  */
-package com.netflix.conductor.dao.es5.index.query.parser;
+package com.netflix.conductor.dao.es.index.query.parser;
 
 import java.io.InputStream;
 import java.math.BigDecimal;
