+++ es5-persistence/src/test/java/com/netflix/conductor/dao/es/index/query/parser/TestConstValue.java
@@ -16,14 +16,9 @@
 /**
  * 
  */
-package com.netflix.conductor.dao.es5.index.query.parser;
+package com.netflix.conductor.dao.es.index.query.parser;
 
 import org.junit.Test;
-
-import com.netflix.conductor.dao.es5.index.query.parser.ConstValue;
-import com.netflix.conductor.dao.es5.index.query.parser.ListConst;
-import com.netflix.conductor.dao.es5.index.query.parser.ParserException;
-import com.netflix.conductor.dao.es5.index.query.parser.Range;
 
 import java.util.List;
