+++ es5-persistence/src/test/java/com/netflix/conductor/dao/es/index/query/parser/TestExpression.java
@@ -16,14 +16,9 @@
 /**
  * 
  */
-package com.netflix.conductor.dao.es5.index.query.parser;
+package com.netflix.conductor.dao.es.index.query.parser;
 
 import org.junit.Test;
-
-import com.netflix.conductor.dao.es5.index.query.parser.ConstValue;
-import com.netflix.conductor.dao.es5.index.query.parser.Expression;
-import com.netflix.conductor.dao.es5.index.query.parser.GroupedExpression;
-import com.netflix.conductor.dao.es5.index.query.parser.NameValue;
 
 import java.io.BufferedInputStream;
 import java.io.ByteArrayInputStream;
